ipq807x: more fixes

+++ feeds/bluetooth-cc2652/cc2652/files/cc2562-reset.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 
-echo 1 > /sys/class/gpio/ble_enable/value
+if [ -f /sys/class/gpio/ble_enable/value ]; then
+	echo 1 > /sys/class/gpio/ble_enable/value
+fi
 echo 1 > /sys/class/gpio/ble_backdoor/value
 echo 1 > /sys/class/gpio/ble_reset/value
 sleep 1
